+++ tests/test_abstract_benchmark.py
@@ -1,12 +1,17 @@
 import pytest
 
-from hpobench.abstract_benchmark import AbstractBenchmark
+from hpobench.abstract_benchmark import AbstractBenchmark, AbstractMultiObjectiveBenchmark
 
-with pytest.raises(NotImplementedError):
-    AbstractBenchmark.get_configuration_space()
 
-with pytest.raises(NotImplementedError):
-    AbstractBenchmark.get_fidelity_space()
+def test_abstract_benchmark():
+    with pytest.raises(NotImplementedError):
+        AbstractBenchmark.get_configuration_space()
 
-with pytest.raises(NotImplementedError):
-    AbstractBenchmark.get_meta_information()
+    with pytest.raises(NotImplementedError):
+        AbstractBenchmark.get_fidelity_space()
+
+    with pytest.raises(NotImplementedError):
+        AbstractBenchmark.get_meta_information()
+
+    with pytest.raises(NotImplementedError):
+        AbstractMultiObjectiveBenchmark.get_objective_names()